1단계 - 지뢰 찾기(그리기) : Field.kt와 FieldInfo.kt 패키지 분리
