+++ src/main/kotlin/domain/cell/Empty.kt
@@ -1,4 +1,4 @@
-package domain.Cell
+package domain.cell
 
 data class Empty(override val row: Int, override val column: Int, override val symbol: String = EMPTY) : Cell {
     companion object {
