+++ src/main/kotlin/domain/cell/Mine.kt
@@ -1,4 +1,4 @@
-package domain.Cell
+package domain.cell
 
 data class Mine(override val row: Int, override val column: Int, override val symbol: String = MINE) : Cell {
     companion object {
