+++ src/main/kotlin/domain/field/FieldInfo.kt
@@ -1,0 +1,4 @@
+package domain.field
+
+data class FieldInfo(val height: Int, val width: Int, val mineCount: Int)
+
